refactor: create separate file FileImportExport and place in new components directory
